Updated quicksync URL & added SNAPSHOT_PRUNING
The Quicksync manifest has moved and the curl call in "run.sh" does not follow HTTP302 redirects.
As a workaround, the new URL is provided.

Additionally, Quicksync offers leveldb and rocksdb snapshots for cryptoorg chain so the default $SNAPSHOT_PRUNING value can not fetch a valid URL.
Setting SNAPSHOT_PRUNING to "leveldb-pruned" resolves this.