run.sh: add snapshot destination in tar cmd
When a SNAPSHOT_URL is specified, the content is stored in $PROJECT_ROOT/data/data.
For example: -e SNAPSHOT_URL=https://snapshots.polkachu.com/snapshots/akash/akash_10400871.tar.lz4
And the earliest and latest block time are "earliest_block_time": "1970-01-01T00:00:00Z".

run.sh is deleting the current data directory, with the destitation in the tar command, the snapshot is stored in $PROJECT_ROOT/data and the block data are correct.