Make COSMOVISOR_ENABLED and START_CMD work as intended (take 3)
With the "CMD $START_CMD" directive ("shell form") docker starts the container
with "$ENTRYPOINT /bin/sh -c $START_CMD". This is bad because it breaks both
cosmovisor and snapshot.sh (the latter has an explicit workaround). It is also
unnecessary because the $START_CMD can (and should) be enforced by `run.sh`.

Remove "CMD $START_CMD", since it is passed in the env and handled anyhow by
`run.sh`. Replace with "CMD []" to un-inherit any "CMD" directives from parent
images.