Fix regression for the case of COSMOVISOR_ENABLED without START_CMD
This fixes a glitch in commit 3aacc3485f8981008f21e3251cc9b2f4300d73df.

When starting without args and without START_CMD, the command for regular
(as-is) and snapshot are "$PREFIX_CMD $PROJECT_BIN start", whereas for
cosmovisor it should be "$PREFIX_CMD start".

Fix using a new DEFAULT_CMD that is set together with PREFIX_CMD based on
the selected mode.